added frontend ui

+++ backend/controllers/goalController.js
@@ -1,10 +1,12 @@
 const asyncHandler = require('express-async-handler')
 const Goal = require('../models/goalModel')
-//@desc get goals
+const User = require('../models/userModel')
+
+//@desc get logged in user's goals
 //@route GET /api/goals
 //@access private
 const getGoals = asyncHandler(async(req,res) =>{
-    const goals = await Goal.find()
+    const goals = await Goal.find({user: req.user.id})
     res.status(200).json(goals)
 })
 
@@ -18,6 +20,7 @@ const setGoal = asyncHandler(async (req, res) => {
     }
     const goal = await Goal.create({
         text: req.body.text,
+        user: req.user.id
     })
     res.status(200).json(goal)
 })
@@ -31,6 +34,18 @@ const updateGoal = asyncHandler(async(req, res) => {
         res.status(400)
         throw new Error('Goal not found.')
     }
+    //get user
+    const user = await User.findById(req.user.id)
+    //check if user exists
+    if(!user){
+        res.status(401)
+        throw new Error('User not found')
+    }
+    //make sure logged in user matches the goal user
+    if(goal.user.toString() !== user.id){
+        res.status(401)
+        throw new Error('User not authorized')
+    }
     const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.body, {new: true})
     res.status(200).json(updatedGoal)
 })
@@ -43,6 +58,18 @@ const deleteGoal = asyncHandler(async (req, res) => {
     if(!goal){
         res.status(400)
         throw new Error('Goal not found.')
+    }
+    //get user
+    const user = await User.findById(req.user.id)
+    //check if user exists
+    if (!user) {
+        res.status(401)
+        throw new Error('User not found')
+    }
+    //make sure logged in user matches the goal user
+    if (goal.user.toString() !== user.id) {
+        res.status(401)
+        throw new Error('User not authorized')
     }
      await Goal.deleteOne(goal)
     res.status(200).json({id: req.params.id})
